+++ src/main/java/roomescape/reservation/ui/AdminReservationRestController.java
@@ -1,0 +1,55 @@
+package roomescape.reservation.ui;
+
+import static roomescape.auth.domain.AuthRole.ADMIN;
+
+import jakarta.validation.Valid;
+import java.util.List;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ModelAttribute;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+import roomescape.auth.domain.RequiresRole;
+import roomescape.reservation.application.ReservationService;
+import roomescape.reservation.ui.dto.request.CreateReservationRequest;
+import roomescape.reservation.ui.dto.request.ReservationsByfilterRequest;
+import roomescape.reservation.ui.dto.response.ReservationResponse;
+
+@RestController
+@RequestMapping("/admin")
+@RequiresRole(authRoles = {ADMIN})
+@RequiredArgsConstructor
+public class AdminReservationRestController {
+
+    private final ReservationService reservationService;
+
+    @PostMapping("/reservations")
+    public ResponseEntity<ReservationResponse> create(
+            @RequestBody @Valid final CreateReservationRequest request
+    ) {
+        final ReservationResponse response = reservationService.create(request);
+
+        return ResponseEntity.status(HttpStatus.CREATED)
+                .body(response);
+    }
+
+    @GetMapping("/reservations")
+    public ResponseEntity<List<ReservationResponse>> findAll() {
+        final List<ReservationResponse> reservationResponses = reservationService.findAll();
+
+        return ResponseEntity.ok(reservationResponses);
+    }
+
+    @GetMapping("/reservations/filtered")
+    public ResponseEntity<List<ReservationResponse>> findAllByFilter(
+            @ModelAttribute @Valid final ReservationsByfilterRequest request
+    ) {
+        final List<ReservationResponse> reservationResponses = reservationService.findAllByFilter(request);
+
+        return ResponseEntity.ok(reservationResponses);
+    }
+}
