+++ src/main/java/roomescape/reservation/ui/ReservationRestController.java
@@ -20,7 +20,6 @@ import roomescape.auth.domain.RequiresRole;
 import roomescape.reservation.application.ReservationService;
 import roomescape.reservation.ui.dto.request.AvailableReservationTimeRequest;
 import roomescape.reservation.ui.dto.request.CreateReservationRequest;
-import roomescape.reservation.ui.dto.request.ReservationsByfilterRequest;
 import roomescape.reservation.ui.dto.response.AvailableReservationTimeResponse;
 import roomescape.reservation.ui.dto.response.ReservationResponse;
 
@@ -30,19 +29,8 @@ public class ReservationRestController {
 
     private final ReservationService reservationService;
 
-    @PostMapping("/admin/reservations")
-    @RequiresRole(authRoles = {ADMIN})
-    public ResponseEntity<ReservationResponse> create(
-            @RequestBody @Valid final CreateReservationRequest request
-    ) {
-        final ReservationResponse response = reservationService.create(request);
-
-        return ResponseEntity.status(HttpStatus.CREATED)
-                .body(response);
-    }
-
     @PostMapping("/reservations")
-    @RequiresRole(authRoles = {MEMBER, ADMIN})
+    @RequiresRole(authRoles = {ADMIN, MEMBER})
     public ResponseEntity<ReservationResponse> create(
             @RequestBody @Valid final CreateReservationRequest.ForMember request,
             final MemberAuthInfo memberAuthInfo
@@ -64,34 +52,19 @@ public class ReservationRestController {
         return ResponseEntity.noContent().build();
     }
 
-    @GetMapping("/admin/reservations")
-    @RequiresRole(authRoles = {ADMIN})
-    public ResponseEntity<List<ReservationResponse>> findAll() {
-        final List<ReservationResponse> reservationResponses = reservationService.findAll();
-
-        return ResponseEntity.ok(reservationResponses);
-    }
-
     @GetMapping("/reservations-mine")
     @RequiresRole(authRoles = {ADMIN, MEMBER})
-    public ResponseEntity<List<ReservationResponse.ForMember>> myReservations(final MemberAuthInfo memberAuthInfo) {
+    public ResponseEntity<List<ReservationResponse.ForMember>> myReservations(
+            final MemberAuthInfo memberAuthInfo
+    ) {
         return ResponseEntity.ok()
                 .body(reservationService.findReservationsByMemberId(memberAuthInfo.id()));
     }
 
-    @GetMapping("/admin/reservations/filtered")
-    @RequiresRole(authRoles = {ADMIN})
-    public ResponseEntity<List<ReservationResponse>> findAllByFilter(
-            @ModelAttribute @Valid final ReservationsByfilterRequest request
-    ) {
-        final List<ReservationResponse> reservationResponses = reservationService.findAllByFilter(request);
-
-        return ResponseEntity.ok(reservationResponses);
-    }
-
     @GetMapping("/reservations/available-times")
     public ResponseEntity<List<AvailableReservationTimeResponse>> findAvailableReservationTimes(
-            @ModelAttribute @Valid final AvailableReservationTimeRequest request) {
+            @ModelAttribute @Valid final AvailableReservationTimeRequest request
+    ) {
         final List<AvailableReservationTimeResponse> availableReservationTimes =
                 reservationService.findAvailableReservationTimes(request);
